requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
apscheduler
yfinance
streamlit
plotly